+++ changelog.md
@@ -22,6 +22,7 @@
 - [actions/create_return_invoice] new action: credit notes @maramalshen
 - Update Error Logs in Actions & Commands @maramalshen
 - Inject Body of the action in the details.meta @maramalshen
+- [actions/create_refund] refund.account_id = rep.integration_meta.qoyod_refund_account_id || options.data.refundAccountId @maramalshen
 
 ### Changed